requirements.txt
django>=4.2
djangorestframework

psycopg2-binary

celery

redis

pandas
openpyxl

python-dotenv

django-cors-headers

django-extensions

gunicorn
